Week 3/writing-app/requirements.txt
gradio>=4.0.0
requests>=2.31.0
Pillow>=10.0.0
openai>=1.0.0
python-dotenv
PyYAML>=6.0.1
manga-ocr>=0.1.8
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.6.0+cpu
torchvision==0.21.0+cpu